+++ CHANGES.rst
@@ -4,6 +4,7 @@ CHANGES
 tip (unreleased)
 ----------------
 
+- added manager_from (thanks George Sakkis)
 - added StatusField, MonitorField, TimeFramedModel, and StatusModel
   (thanks Jannis Leidel)
 - deprecated ChoiceEnum and replaced with Choices